requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.2.6
django-bootstrap3==11.1.0
djangorestframework==3.10.3
future==0.18.1
idna==2.8
lxml==4.4.2
Pillow==6.2.1
pycryptodome==3.9.4
PyMySQL==0.9.3
pytz==2019.3
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.7
